Cambia lugar btns Categorias a Sidebar
Y arregla Links de Cards a /juira/:id pero todavia usando datos hardcode
